pyproject.toml
[tool.poetry]
name = "llamaindex-sql-agents"
version = "0.1.0"
description = ""
authors = ["Sérgio Pereira da Silva <sergio.pds@outlook.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, < 3.13"
llama-index-llms-openai = "^0.1.25"
llama-index-embeddings-openai = "^0.1.10"
SQLAlchemy = "^2.0.31"
openai = "^1.35.10"
python-dotenv = "^1.0.1"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
pydantic = "^2.8.2"
fastapi = "^0.111.0"
streamlit = "^1.36.0"
qdrant-client = "^1.10.0"
llama-index = "^0.10.52"
llama-index-vector-stores-qdrant = "^0.2.11"
arize-phoenix = {extras = ["evals"], version = "^4.7.1"}
llama-index-callbacks-arize-phoenix = "^0.1.6"

[tool.poetry.group.dev.dependencies]
openpyxl = "^3.1.5"
pandas = "^2.2.2"
ipykernel = "^6.29.5"
jupyter = "^1.0.0"
ipywidgets = "^8.1.3"
pyvis = "^0.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
